+++ crates/header-translator/src/id.rs
@@ -238,10 +238,14 @@ impl<N: ToOptionString> ItemIdentifier<N> {
     }
 
     pub fn with_name(name: N, entity: &Entity<'_>, context: &Context<'_>) -> Self {
-        let location = context.get_location(entity).unwrap_or_else(|| {
+        let mut location = context.get_location(entity).unwrap_or_else(|| {
             error!(?entity, "ItemIdentifier from unknown header");
             Location::from_components(vec!["__Unknown__".into()])
         });
+
+        if let Some("IOSurfaceRef") = name.to_option() {
+            location = Location::from_components(vec!["IOSurface".into(), "IOSurfaceRef".into()]);
+        }
 
         Self { name, location }
     }