+++ crates/header-translator/src/main.rs
@@ -136,7 +136,10 @@ fn main() -> Result<(), BoxError> {
         let mut includes_file = fs::File::create(&includes).unwrap();
         // Make sure that we pick the IOSurfaceRef that IOSurface defines,
         // instead of the one that CoreGraphics defines.
-        writeln!(&mut includes_file, "#import <IOSurface/IOSurface.h>")?;
+        if platform_config_filter(&config.libraries["IOSurface"]) {
+            writeln!(&mut includes_file, "#include <IOSurface/IOSurfaceRef.h>")?;
+            writeln!(&mut includes_file, "#include <IOSurface/IOSurfaceObjC.h>")?;
+        }
         // Part of .modulemap
         writeln!(&mut includes_file, "#import <CoreFoundation/CFPluginCOM.h>")?;
         for lib in config.libraries.values() {