+++ framework-crates/objc2-io-surface/src/lib.rs
@@ -24,3 +24,7 @@ pub use self::generated::*;
 pub(crate) type Boolean = u8;
 #[allow(dead_code)]
 pub(crate) type OSType = u32;
+
+/// [Apple's documentation](https://developer.apple.com/documentation/iosurface/iosurfaceref?language=objc)
+#[cfg(feature = "IOSurfaceRef")]
+pub type IOSurfaceRef = *mut core::ffi::c_void;